Fix comments by nnatter
