+++ src/Sulu/Bundle/AdminBundle/Resources/config/services.xml
@@ -167,6 +167,13 @@
             <tag name="console.command"/>
         </service>
 
+        <service id="sulu_admin.update_build_command" class="Sulu\Bundle\AdminBundle\Command\UpdateBuildCommand">
+            <argument type="service" id="http_client" />
+            <argument>%kernel.project_dir%</argument>
+            <argument>%sulu.version%</argument>
+            <tag name="console.command"/>
+        </service>
+
         <service
             id="sulu_admin.download_language_command"
             class="Sulu\Bundle\AdminBundle\Command\DownloadLanguageCommand"
